Added a simple filter for removing duplicate URLs.
It checks each URL to determine if there is another, shorter URL
already found and that is working. If so the longer URL is removed.

This simple approach will work in most cases but it fails for this
example:
	http://example.com
	http://example.com.tw

In this case the second URL would get removed even though the two URLs
are probably not for the same cache.